#24657 Fixed bug causing hierarchical data models to potentially crash.
This seems to be caused by a known issue in Shiboken's object release
behaviour, where it can end up in an endless recursion in some cases
where there is a double bind between a parent and a child (see 
https://bugreports.qt-project.org/browse/PYSIDE-158). This change works
around this shortcoming by always explicitly clearing the model in a
depth first recursion, starting with children and going up,
thereby avoiding the errouneous shiboken state.

A side effect of this change is that the model now emits more signals
upon destruction and upon data refresh - because items are now deleted
one by one, it may trigger for example selection changed signals in the
view as a result - stuff that was previously not triggered since the
deletion happened as a single operation.