#28557 #28558 Bug fixes for clearing model. Better disk structure.
Improved workarounds for a known pyside/qt issue where clearing and
resetting of models will result in dangling pointers on the PySide side.
This change implements close() and reset() for the Shotgun Model in the
interest of stability - anyone trying to call reset() will now get an
exception raised instead of a (highly likely) scenario with stability
issues or crashes. close() is reimplemented in order to avoid the
default C++ implementation and instead use the hand crafted recursion
based clearing that ensures that all objects both on the C++ and Python
side are removed.

Thumbnail and cached query paths are restructured to better handle long
path name as we were running into issues with MAX_PATH on windows. We
had to deal with taking an S3 guid and shard that into the file system
(which wasn't done nicely before) so now all thumbnail paths that
shotgun generate are first hashed and then sharded on disk. This should
result in an even distribution of thumbnails on disk.

Paths logic for shotgun caches have been updated as well in order to
avoid MAX_PATH on windows (260 char limitation). Warnings are emitted if
the paths are getting close.

Closes #12