#30748 Improved cache logic in shotgun model
This tweaks the messaging and logic around how caches are structured and
handled. This change changes the logic for how caches are laid out on
disk and old cache files will no longer be reachable. The change
addresses several advanced use cases:

- The thumbnail flag is now part of the cache hashing, meaning that
shotgun models with and without thumbnails will cache into different
locations on disk. This addresses a known issue where the sg model is
run in a thread. previously, a QPixmap: It is not safe to use pixmaps
outside the GUI thread would be reported from the thread in the case the
implementation hit a cache which had thumbnails stored. By separating
the cache based on the thumb flag, caches are guaranteed to be 'clean'
from thumb data whenever the no thumbs flag is used.
- The schema_generation parameter is now cached with the path rather
than previously with the file content. This fixes a subtle inconvenience
where previously, if you had two versions of an app running, these could
"compete" for a cache file, one writing it out only for it to be
overwritten by another. The situation would also arise if you happened
to have two different apps calling the same query with same params, but
having two schema generation markers. Both these scenarios are
plausible.
- The seed parameter is a runtime, data centric parameter and has been
moved down into the "runtime" portion of the hash, resulting in cleaner
hash file layout when using seed based hashing.
- Improved doc strings.
- Improved exceptions.

Closes #20