For #35801, file save freeze in workfiles 2
Signalling between two different threads in PySide is broken in several versions of PySide. There are very subtle race conditions that arise when there is a lot of signalling between two threads. Some of these things have been fixed in later versions of PySide, but most hosts integrate PySide 1.2.2 and lower, which are victim of this race condition.

The background task manager does a lot on inter-threads communications and therefore can easily fall pray to these deadlocks that exist within PySide.

Therefore, the background task manager now uses a polling system to process results from the worker threads. The worker threads inserts results inside a queue and the background manager's thread will flush the queue at 100 milliseconds intervals.

Also note that because there is a 100 milliseconds interval between two polls, it means that the will automatically be a delay between the end of a task and the beginning of another task.

Closes https://github.com/shotgunsoftware/tk-framework-shotgunutils/pull/33