#37113 Shotgun Hierarchy Model
These changes introduce a new ShotgunHierarchyModel which retrieves data via
the new shotgun python-api call nav_expand(). The structure and organization of
the model is very similar to the existing ShotgunModel and in fact they now
share a new base class called ShotgunQueryModel.

Added a SimpleShotgunModel class that inherits ShotgunHierarchyModel that
allows for very simple creation of a hierarchy model.

Closes #49.