#38172 Allows for graceful failure when Hierarchy API is not available.
If you create a ShotgunHierarchyModel and the shotgun connection doesn't
support the nav_expand() query, a single item will be added to the model that
says the version of SG must be v7.0.2 in order to use the model. A warning is
also logged.

Adds a method: shotgun_model.hierarchy_is_supported() to allow graceful
fallback for clients.

Closes #51.