For #40080, Added methods to retrieve thumbnail source files (#59)
Added `ShotgunDataRetriever.download_thumbnail_source(entity_type, entity_id, bundle)` and `ShotgunDataRetriever.request_thumbnail_source(entity_type, entity_id)`, convenience and compatibility methods for quick and easy synchronous (download_thumbnail_source) and asynchronous (request_thumbnail_source) thumbnail download. This methods will retrieve the source file for a thumbnail given a shotgun entity type and id. If the resolved thumbnail source file has already been cached, a path to it will be returned instantly. Otherwise, it will be downloaded from Shotgun and placed in the standard cache location on disk. The full path to cached thumbnail is returned. The returned thumbnail file in the original format and resolution it was uploaded to Shotgun as, which should be considered arbitrary.  Closes #59.