#40080, Remove legacy cached .jpeg thumbnails (#60)
In ShotgunDataRetriever._get_thumbnail_path(), if more than one cache file is found for a Shotgun thumbnail URL, check to see if one of them is the legacy ".jpeg" file.  If so, remove it from the list of cached files found as well as from disk. Once removed, if there is still more than one cache file, log a warning and return the first match found.