#46914 caching delete bug (#85)
Fixes a bug which was causing deletion of pipeline configurations to go un-noticed by the update detection. With this change, the hash representing the current state of the software entity + pipeline configuration entity is computed as the updated_at timestamps for all records in the software entity + all pipeline configurations belonging to projects which have not be archived. (this constraint is not not have an ever growing number of configurations to consider of the life cycle of a shotgun site).