SG-3019: Adds tests for external_config! (#102)
This adds test coverage to the top level of the external_config API, which will also exercise some of the underlying logic. It is by no means complete coverage, but it's a good start, and it hits the portions of the API that can easily be called synchronously.

This covers:

- Getting PC entities from Shotgun.
- Processing of PC entities from Shotgun into ExternalConfiguration objects.
- Failure to retrieve a list of PC entities from Shotgun.

The second item on the above list also exercises the basic interface of the ExternalConfiguration object.