SG-18486 Now suppresses warnings that were popping up in browser integration. (#119)
Suppresses the imp import depreciation warnings, and the libpng warnings from QT. These warnings were appearing as a popup in the browser on the Shotgun web page when running a toolkit action.

The imp deprecation warnings are only raised when the script that is directly executed by the python contains an imp import. We needed to handle it here because external config module of SG utils executes this script in a new python instance.