Upgrading to PySide2 5.15 causes a deadlock when the WorkerThread's shut_down method is called.
In the past, we've had issues when mixing Qt and Python threading primitives and this is indeed what is going on here. The WorkerThread is a QThread but we use Python's Mutex and Condition classes to synchronize threads. I've simply modified the WorkerThread class to derive from Python's Thread class and now everything is fine.